Refactor: userId 기반 API를 세션 기반 /me 구조로 리팩토링

+++ backend/src/main/java/com/ssafy/chillandcode/model/dao/UserDao.java
@@ -11,13 +11,13 @@ public interface UserDao {
 	int insertUser(User user);
 	
 	//회원 정보 조회
-	User selectUser(int userId);
+	User selectUser(long userId);
 	
 	//회원 정보 수정
 	int updateUser(User user);
 	
 	//회원 정보 삭제(탈퇴)
-	int softDelete(int userId);
+	int softDelete(long userId);
 	
 	//로그인
 	User login(User user);
