+++ backend/src/main/java/com/ssafy/chillandcode/model/service/UserService.java
@@ -7,13 +7,13 @@ public interface UserService {
 	boolean insertUser(User user);
 
 	// 회원 정보 조회
-	User selectUser(int userId);
+	User selectUser(long userId);
 
 	// 회원 정보 수정
 	boolean updateUser(User user);
 
 	// 회원 정보 삭제(탈퇴)
-	boolean softDelete(int userId);
+	boolean softDelete(long userId);
 
 	// 로그인
 	User login(User user);
